+++ src/main/java/com/github/bezsias/multimap/Util.java
@@ -14,12 +14,12 @@ class Util {
             to.write(buf, 0, len);
     }
 
-    static void writeShort(short value, byte[] array, int pos) {
+    static void writeShort(int value, byte[] array, int pos) {
         array[pos] = (byte)((value >> 8) & 0xFF);
         array[pos+1] = (byte)(value & 0xFF);
     }
 
-    static short readShort(byte[] array, int pos) {
-        return (short)((array[pos] & 0xFF) << 8 | (array[pos+1] & 0xFF));
+    static int readShort(byte[] array, int pos) {
+        return ((array[pos] & 0xFF) << 8 | (array[pos+1] & 0xFF));
     }
 }
